Fibonacci series

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int i = 0;
-	unsigned long long fibo[50];
+	long fibo[50];
 
 	fibo[0] = 1;
 	fibo[1] = 2;
@@ -17,7 +17,7 @@ int main(void)
 	for (i = 0; i <= 49; i++)
 	{
 		fibo[i + 2] = fibo[i]  + fibo[i + 1];
-		printf("%llu, ", fibo[i]);
+		printf("%ld, ", fibo[i]);
 	}
 	printf("\n");
 	return (0);
